moved the serrver and package json and git ignore files to the root of the repository
