--- .env/package.json
@@ -1,30 +1,0 @@
-{
-    "name": "crowd-funding",
-    "version": "1.0.0",
-    "description": "",
-    "main": "server.js",
-    "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1",
-      "start": "node server.js",
-      "seed": "node seeds/seed.js"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "MIT",
-    "dependencies": {
-      "bcrypt": "^5.0.0",
-      "connect-session-sequelize": "^7.0.4",
-      "dotenv": "^8.2.0",
-      "express": "^4.17.1",
-      "express-handlebars": "^5.2.0",
-      "express-session": "^1.17.1",
-      "pg": "^8.11.3",
-      "sequelize": "^6.3.5"
-    },
-    "devDependencies": {
-      "eslint": "^7.12.1",
-      "eslint-config-prettier": "^6.15.0",
-      "prettier": "^2.1.2"
-    }
-  }
-
